Rename EmbeddedResourceDocumentViewer to Document
