+++ apps/charterafrica/src/components/Document/index.js
@@ -1,0 +1,3 @@
+import Document from "./Document";
+
+export default Document;
