--- apps/charterafrica/src/components/EmbeddedResourceDocumentViewer/index.js
@@ -1,3 +1,0 @@
-import EmbeddedResourceDocumentViewer from "./EmbeddedResourceDocumentViewer";
-
-export default EmbeddedResourceDocumentViewer;
