+++ apps/charterafrica/src/pages/[...slugs].page.js
@@ -6,10 +6,10 @@ import CommunityPlatforms from "@/charterafrica/components/CommunityPlatforms";
 import Consultations from "@/charterafrica/components/Consultations";
 import Dataset from "@/charterafrica/components/Dataset";
 import Datasets from "@/charterafrica/components/Datasets";
+import Document from "@/charterafrica/components/Document";
 import Documents from "@/charterafrica/components/Documents";
 import Ecosystem from "@/charterafrica/components/Ecosystem";
 import EmbeddedDocumentViewer from "@/charterafrica/components/EmbeddedDocumentViewer";
-import EmbeddedResourceDocumentViewer from "@/charterafrica/components/EmbeddedResourceDocumentViewer";
 import Entity from "@/charterafrica/components/Entity";
 import Explainers from "@/charterafrica/components/Explainers";
 import FAQ from "@/charterafrica/components/FAQ";
@@ -46,7 +46,7 @@ const componentsBySlugs = {
   ecosystem: Ecosystem,
   "embedded-playlist": Consultations,
   "embedded-document-viewer": EmbeddedDocumentViewer,
-  "embedded-resource-document-viewer": EmbeddedResourceDocumentViewer,
+  document: Document,
   explainers: Explainers,
   faq: FAQ,
   "featured-post": FeaturedPostCard,
